Split express code into watson and alexa
